allow multiple monit rc files

+++ cli/lib/cli/job_builder.rb
@@ -88,7 +88,8 @@ module Bosh::Cli
         raise InvalidJob, "There are unused template files for job '#{name}': %s" % [ extra_templates.join(", ")]
       end
 
-      unless File.exists?(File.join(job_dir, "monit"))
+      # should we only allow either monit || foo.monit
+      unless File.exists?(File.join(job_dir, "monit")) || Dir.glob("#{job_dir}/*.monit").size > 0
         raise InvalidJob, "Cannot find monit file for '#{name}'"
       end
 
@@ -117,9 +118,18 @@ module Bosh::Cli
         copied += 1
       end
 
-      FileUtils.cp(File.join(job_dir, "monit"), build_dir, :preserve => true)
+      Dir.glob("#{job_dir}/*.monit").each do |file|
+        FileUtils.cp(file, build_dir, :preserve => true)
+        copied += 1
+      end
+
+      if File.exist?(File.join(job_dir, "monit"))
+        FileUtils.cp(File.join(job_dir, "monit"), build_dir, :preserve => true)
+        copied += 1
+      end
+
       FileUtils.cp(File.join(job_dir, "spec"), File.join(build_dir, "job.MF"), :preserve => true)
-      copied += 2
+      copied += 1
       copied
     end
